backend/requirements.txt
coreapi
Django==1.11.8
django-debug-toolbar
django-filter
djangorestframework
flake8
Markdown
psycopg2
envparse
celery
django_celery_results
django-celery-beat
django-cas-ng
pyldap
cryptography
requests
django-jsoneditor
gunicorn
django-pyodbc-azure
python-jose
